requirements.txt
openai-whisper
pyperclip
torch
